requirements.txt
python-dateutil
bs4
requests
selenium